requirements.txt
pylint==2.3.1
pytest==5.0.0
webdriver-manager==1.8.2
selenium==3.141.0
allure-pytest
allure-python-commons